refactor(cockpit): move review plan tests to owner module
